requirements.txt
# Core dependencies for file format conversion

# PDF handling
PyPDF2>=3.0.1
reportlab>=4.0.0

# Microsoft Office formats
python-docx>=0.8.11
python-pptx>=0.6.21

# E-book formats
ebooklib>=0.18

# Additional utilities
typing-extensions>=4.0.0
